Memory Leak Fixes
This fix prevents large scope and element leaks when removing an ng-grid.
The primary issue was that when the default templates were compiled empty
text nodes, and for the row & header row templates an ng-repeat comment
node, were given access to the current scope via an entry in the jQuery
data cache. For some reason even when those elements were removed their
associated entries in the data cached weren't cleared out leading to a
scope memory leak.

Due to the how interconnected scopes are this one scope leak caused a lot
of other scopes and their associated elements to be leaked as well.

I've updated the build tools to minify the templates,
which should've been done anyways, and this prevents most of the leak. I also
had to wrap the default row & header row templates in an empty div to
avoid the first ng-repeat comment from being associated with a scope.

Additionally I added some cleanup code to clear out jQuery references and event listeners that were being left around.