Fix scrolling behavior
Fix scrolling behavior to work like a grid should. When using the keyboard
to move selection the grid will be scrolled to ensure the row is
completely visible. Also added support to scrolling to the newly selected
row when using the kyboard to move selection if the newly selected row is
completely off screen.

As part of this the lastSelectedRow property on the selection provider
will always be an original ng-row object and never a clone. This is
needed because when scrolling to a row that is completely off screen if
the lastSelectedRow was a clone it didn't refer to the actual row that was
selected last.

I also commented out the unusual behavior of focusing the grid on
mousewheel as it was breaking scrolling.

Additionally I updated the version of grunt-contrib-uglify to be new
enough to have a fixed version of uglify to ensure the build is
successful.